基础功能已完善，需要做个首页，还有其他功能的侧边栏，最后加入rbac。物料申请还没做。

+++ src/main/java/xyz/kakusei/fams/mapper/IMaterialTypeMapper.java
@@ -5,9 +5,9 @@ import xyz.kakusei.fams.entity.MaterialType;
 import java.util.List;
 
 public interface IMaterialTypeMapper {
-    MaterialType queryById(Byte id);
+    MaterialType queryById(Integer id);
     List<MaterialType> queryAll();
     void insert(MaterialType materialType);
     void update(MaterialType materialType);
-    void deleteById(Byte id);
+    void deleteById(Integer id);
 }
